Make it compile for my eCos BlackFin checkout. - numerous casts because "%zd" ever requires a size_t, even for a manifest   constant; and because a uint32_t should be cast to a long unsigned to be   printed with "%lu" (long unsigned is at least 32-bit) - move include/urjtag/urjtag.h to include/urjtag/urjtag.h.in and only   export .h files for modules actually configured: USB, SVF, BSDL are   only conditionally exported
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@1601 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ urjtag/include/urjtag/cmd.h
@@ -59,6 +59,6 @@ char *urj_cmd_find_next (const char *text, int state);
 /* @@@@ RFHH candidate to become local in src/cmd/ after cable refactor */
 int urj_cmd_params (char *params[]);
 /* @@@@ RFHH candidate to become local in src/cmd/ after cable refactor */
-int urj_cmd_get_number (const char *s, unsigned int *i);
+int urj_cmd_get_number (const char *s, long unsigned *i);
 
 #endif /* URJ_CMD_H */
